backend/app/pyproject.toml
[tool.poetry]
name = "fastapireact"
version = "0.1.0"
description = ""
authors = ["pratik <prateeklilhare@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
uvicorn = "^0.18.1"
gunicorn = "^20.1.0"
aioredis = "^2.0.1"
aiohttp = "^3.8.1"
tortoise-orm = "^0.19.2"
asyncpg = "^0.26.0"
aerich = "^0.6.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
